+++ spec/cucumber/cli/options_spec.rb
@@ -272,31 +272,31 @@ module Cucumber
           end
 
           it 'includes any non-STDOUT formatters from the profile' do
-            given_cucumber_yml_defined_as({'html' => %w[--format html -o features.html]})
-            options.parse!(%w{--format progress --profile html})
+            given_cucumber_yml_defined_as({'json' => %w[--format json -o features.json]})
+            options.parse!(%w{--format progress --profile json})
 
-            expect(options[:formats]).to eq [['progress', {}, output_stream], ['html', {}, 'features.html']]
+            expect(options[:formats]).to eq [['progress', {}, output_stream], ['json', {}, 'features.json']]
           end
 
           it 'does not include STDOUT formatters from the profile if there is a STDOUT formatter in command line' do
-            given_cucumber_yml_defined_as({'html' => %w[--format html -o features.html --format pretty]})
-            options.parse!(%w{--format progress --profile html})
+            given_cucumber_yml_defined_as({'json' => %w[--format json -o features.json --format pretty]})
+            options.parse!(%w{--format progress --profile json})
 
-            expect(options[:formats]).to eq [['progress', {}, output_stream], ['html', {}, 'features.html']]
+            expect(options[:formats]).to eq [['progress', {}, output_stream], ['json', {}, 'features.json']]
           end
 
           it 'includes any STDOUT formatters from the profile if no STDOUT formatter was specified in command line' do
-            given_cucumber_yml_defined_as({'html' => %w[--format html]})
-            options.parse!(%w{--format rerun -o rerun.txt --profile html})
+            given_cucumber_yml_defined_as({'json' => %w[--format json]})
+            options.parse!(%w{--format rerun -o rerun.txt --profile json})
 
-            expect(options[:formats]).to eq [['html', {}, output_stream], ['rerun', {}, 'rerun.txt']]
+            expect(options[:formats]).to eq [['json', {}, output_stream], ['rerun', {}, 'rerun.txt']]
           end
 
           it 'assumes all of the formatters defined in the profile when none are specified on cmd line' do
-            given_cucumber_yml_defined_as({'html' => %w[--format progress --format html -o features.html]})
-            options.parse!(%w{--profile html})
+            given_cucumber_yml_defined_as({'json' => %w[--format progress --format json -o features.json]})
+            options.parse!(%w{--profile json})
 
-            expect(options[:formats]).to eq [['progress', {}, output_stream], ['html', {}, 'features.html']]
+            expect(options[:formats]).to eq [['progress', {}, output_stream], ['json', {}, 'features.json']]
           end
 
           it 'only reads cucumber.yml once' do